refactor(web): move the forced-tools store in with the rest of tools
It sat in `lib/hooks`, which the frontend standards keep for hooks with no
feature home. This one has a home: it holds the tool the next message will
be made to use, and `lib/tools/hooks.ts` already holds the MCP server
reads.

It stays a zustand store. All five consumers call it directly and none
receive it as a prop, which is what a module-level store is for — one
forced tool per app, reachable anywhere, no provider to mount.

Also collapses ToolsPopover's two separate imports from
`lib/tools/constants` into one.
